change(ffi): move the utd hook from the sync service to the client
