+++ bindings/matrix-sdk-ffi/src/lib.rs
@@ -31,6 +31,7 @@ mod sync_service;
 mod task_handle;
 mod timeline;
 mod tracing;
+mod utd;
 mod utils;
 mod widget;
